# Optimize the connected components labeling algorithm