#### Changes
<!-- Please include a summary of the change and which issue is fixed. -->
I tried to utilize connected components labeling algorithm (CCL) on a GPU, it was too slow for real-time processing. For that reason, I optimized the performance of CCL implementation.

The previous method basically utilizes max pooling to perform watershed. To exclude the "background" it masks the output from max pooling each iteration. Masking is performed using boolean indexing, which is suboptimal for memory accessing.

To improve the performance I used element-wise multiplication instead of indexing. 

I measured the performance on an `602x602` image with `num_iterations=150`. Measured time was computed by taking average from 100 runs:

#### Tested hardware
CPU: AMD Ryzen 7 5800H
GPU: GeForce RTX 3070 Mobile / Max-Q

#### Benchmark
| **Method/Device**           	| **CPU** 	| **GPU** 	|
|-----------------------------	|---------	|---------	|
| **Original implementation** 	|  129.77 ms 	|  34.43 ms 	|
| **New implementation**      	|  67.55 ms 	|   5.07 ms 	|

<!-- Please also include relevant motivation and context. -->
<!-- List any dependencies that are required for this change. -->

Fixes # (issue)

#### Type of change
<!-- Please delete options that are not relevant. -->
- [ ] 📚  Documentation Update
- [x] 🧪 Tests Cases
- [ ] 🐞 Bug fix (non-breaking change which fixes an issue)
- [x] 🔬 New feature (non-breaking change which adds functionality)
- [ ] 🚨 Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] 📝 This change requires a documentation update


#### Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] Did you update CHANGELOG in case of a major change?